Move textmeasure into lib
It's not a d2renderer.
