+++ lib/textmeasure/NOTICE.txt
