+++ lib/textmeasure/rect.go
